Move selectByQuery to select-new
